Exit the Python process on timeout
This is a more effective way of cleaning up active GDAL threads that
would otherwise resume processing on subsequent requests when the
process is thawed.
